flask_backend/requirements.txt
flask
psycopg2-binary
pandas
numpy
scikit-learn
joblib
openai
python-dotenv
